Updated the StringBuilder.
Methods that can fail now return error codes.
Methods now take null terminated strings. Methods that take the length
of the string have been prefixed with 'n', e.g. nappend.
Added to the indexOf methods an option to choose from which index to
search.
Added an init method for initializing a StringBuilder without new().
Added a destroy method for StringBuilder that were not made with new.